Move most Rc tests to alloctests
